src/requirements.txt
fastapi
uvicorn[standard]
gunicorn
slowapi
pyjwt
python-multipart
python-jose[cryptography]
passlib[bcrypt]
langchain[llms]
azure-storage-blob
azure-keyvault-secrets
azure-identity
langchain-google-genai
langchain-openai
langchain-community
langchain.output_parsers.json
langchain_core.output_parsers
langchain_core.runnables
langchain_core.prompts
langchain_core.output_parsers
langchain_core.runnables
langchain_core.prompts

